requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.3
Babel==2.4.0
billiard==3.5.0.2
celery==4.0.2
Django==1.11
django-rest-swagger==0.3.10
djangorestframework==3.6.2
flower==0.9.1
kombu==4.0.2
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
PyYAML==3.12
requests==2.13.0
six==1.10.0
tornado==4.2
vine==1.1.3
